CRAYSAT-42: Update versioning to meet new DST standard
Update the versioning of our RPM to meet the new DST versioning
standard. The current version of our rpm was 0.1.1, so update the
version in the spec file to 0.1.2 to avoid going backwards. The release
contains the timestamp of the build, an underscore, and the short git
commit hash.

See the "DST Versioning Standard" page in confluence for more
information.

Test Description:
Pushed branch and verified that the cray-sat rpm built and had a version
and release matching the standard.