CRAYSAT-29: Create ansible role in a subpackage
Create a crayctldeploy subpackage that contains the ansible role to
be installed in the appropriate ansible_framework directory. Creating it
as a subpackage allows it to be very closely integrated with the code it
is installing and allows it to be built without creating a whole new
Jenkinsfile and multibranch pipeline.

Update the minor version of the package for the added ansible role
functionality.

Test Description:
Pushed and verified that the new subpackage was built by the existing
Jenkinsfile in the multbranch pipeline. Downloaded the resulting package
from artifactory and verified that it contains the expected files.

Tested installation of cray-sat-crayctldeploy package on spottedcow and
execution of the ansible role as follows:

* Added a repo pointing to the artifactory location for this branch
  using `zypper addrepo`.
* Installed `cray-sat-crayctldeploy` using `zypper install`.
* Set `ANSIBLE_ROLES_PATH=/opt/cray/crayctl/ansible_framework/roles/` in
  environment.
* Ran a playbook that included the `cray_sat` role and verified that it
  installed `cray-sat`.
* Ran the playbook again to ensure that it made no changes when the
  latest version of the package was already installed.