CRAYSAT-53: Move jenkins notifications to a separate slack channel
Move jenkins notifications to a new sat-ci channel in slack that is
separate from the shasta_admin_toolkit channel, which will be reserved
for actual conversations between humans.

Test Description:
Pushed branch and verified that the pipeline built successfully and
posted notifications to the sat-ci slack channel and not to the
shasta_admin_toolkit slack channel.