CRAYSAT-86: Add hwinv subcommand and its options to parser
Add the `sat hwinv` subcommand to the SAT parent parser, and add all of
its options, approximately as defined in the "sat hwinv Design"
confluence page. The main difference is that I was able to get by with
just a `--show-<component>-xnames' option to both enable and disable
printing of xnames in summary output.

Add a sat.hwinv.main module that implements a function that does nothing
except log the args it received at debug level and then log an error
message that it is not implemented before exiting with non-zero status.

Add a dependency on a python package called `inflect`. This is used to
generate plural nouns from singular nouns, which is used by the code
that adds arguments to the parser in a consistent manner for the various
component types that can be listed or summarized.

Refactor the parser for each subcommand into its own module in the
subpackage for the subcommand since the one `parser` file was getting
pretty large. This will also probably make it easier to move to an
architecture where subcommands are automatically discovered and made a
part of the program without manually adding them to sat.main in the
future.

Test Description:
Ran in a virtual env with various combinations of options and
`--loglevel debug` to see how the options were set with the debug
message that is printed.