CRAYSAT-78: Filter status output by nodes
Added command line arguments -x, --xnames and -n, --nids that filter the list
of nodes to show only those specified by the given xname or nid, which may be
a single node or a comma separated list.

Test Description:

Code is tested against mock data from the user and the Shasta API. Notable
edge cases such as empty results are checked, as well as samples
representative of normal usage. Error conditions are verified.