CRAYSAT-47: Implement a config file and configure logging
Implement a configuration file in INI format called `sat.ini`, and use
the configuration values within it to configure logging. Install the
configuration file as a config file in the spec file. Implement logic
that verifies the config file for validity and uses defaults that are
hard-coded in the code if options are not present in the config file.

Refactor logging implementation at the same time. Add handlers to the
root logger in functions in the newly created `sat.logging` module. Call
these functions in the main method.In other modules, use
`logging.getLogger(__name__)` to get a logger for that module. This is
the way that python logging docs recommend setting up logging. It will
include the name of the module logging the message in all log messages.

Also add unit tests for both config and logging modules.

Test Description:
All unit tests pass. Installed rpm on surly1-sms1 and verified that the
log directory was created, and the default config file was installed.
Also verified that old configuration files are not replaced if they have
been modified on the system when upgrading the package.

Tested functionality of `SAT_CONFIG_FILE` environment variable for
loading an alternative config file. Tested that warning and error
messages are logged for invalid config files. Broke `/etc/os-release`
and verified that `sat showrev` logged errors both to the file and to
stderr.