CRAYSAT-46: Man pages
Added man pages for 'sat', 'sat showrev', and 'sat cablecheck'.

Test description:

I grabbed the rpm built by jenkins, copied it over to spottedcow-sms1
and I unpacked it in a local directory to verify the internal location
of man pages, as well as the format (.gz) of each man page.