CRAYSAT-47: Move to /var/log/cray/sat.log and ensure dir exists
Move the log file from `/var/log/sat/sat.log` to `/var/log/cray/sat.log`
and ensure that the parent directory exists when creating the log file
handler, and don't create that handler if we can't create that directory
or it doesn't already exist. Add unit tests for this case, and remove
some values from test config files that do not need to be included for
the test cases.

Test Description:
All unit tests pass. Will test manually on surly1-sms1 by installing
rpm, running `sat showrev` with a missing `/etc/os-release` to force
logging of an error, and verify that the error is logged to the file.
Configure sat to log to a file in a directory that doesn't exist, and
then run `sat showrev` with errors. Verify that the directory is
created. Configure sat to log to a file whose parent directory can't be
created because it is a file, and verify that `sat` logs an error about
it and continues on its way without logging to a file.