CRAYSAT-112: Add `sat diag` subcommand.
This commit adds the full implementation of the `sat diag` subcommand.
This command allows manufacturing to peform diagnostics on Rosetta
switches at L1 using the Redfish API through hmjtd.

Users can specify a list of xnames for switches which should be tested
either specifically as command line arguments, or in a newline-delimited
file. The user should also specify a dignostic command with arguments
which should be run. The `sat diag` tool will then run the command on
the switches and poll them until all diagnostics have completed.
The tool will then print the resulting output either to stdout or to
individual files for each switch. A summary of the results is printed.

Test Description:
1. A handful of unit tests are provided, which can be run with
   nosetests.
2. On shandy the following tests were run to ensure behavior worked
   properly:
      - sat diag -f xnames.txt --disruptive -i 1 runDiagPost -q
      - sat diag -x x5192c3r4b0 runMemTester -h
      - sat diag -x x5192c3r4b0 -t 0 runMemTester -h
      - sat diag -x x5192c3r4b0 runMemTester --does-not-exit
      - sat diag -x x5192c3r4b0 fooBarBaz --does-not-exist
      - sat diag -x xNotAnXname fooBarBaz --does-not-exist
      - sat diag runMemTester -h
   The goals of these tests are to test supplying xnames on the
   command line as well as in a file, to test graceful failure when
   bad commands or arguments are passed, to test when bad arguments
   are passed to `sat diag`, and finally to test a normal use-case.