CRAYSAT-48: Query HSM Hardware inventory
Refactor access to the API gateway into its own apiclient module, and
create a class called APIGatewayClient that will abstract away access to
the API gateway. Hopefully this will be useful for implementing
authentication to the API gateway in one place. If not, it can be
refactored. Currently, access to individual APIs can be done by
subclassing the APIGatewayClient and setting the `base_resource_path`.
An `HSMClient` class is created for ease of making requests to the HSM
API.

Also add an option for setting the API gateway host in the config file.

Update sat.status.main to use the new HSMClient class, and pull the API
request out of the `make_raw_table` function and instead pass in a list
of components. This is a better separation of concerns, and it makes
`make_raw_table` a little easier to unit test. Update unit tests
accordingly.

Also update `sat.hwinv.main` to make a query to the HSM hardware
inventory, and just print a table of the nodes for now as a placeholder
for real implementation of all output options.

Other minor fixes:
* Clean up docstrings in sat.status.main to conform to Google style
* Fix camelCase vars in sat.status.main
* Fix pycodestyle violations.

Test Description:
Installed in a virtualenv on spottedcow. Ran `sat status` and `sat
hwinv` commands and verified that output was correct. Ran `sat status`
with some filters and sort options as well. All unit tests pass.