CRAYSAT-138: Unittest interconnects
Unittests for the get_interconnects function within showrev --system.
Verifies that it returns a unique list of interconnect types.

This also modifies the return value from get_interconnects to return a
list, and refactors the function somewhat to move the api-reliant logic
into its own function. This makes it possible to mock the API-centric
parts of this feature, which allows unit testing.

Test description:

All unit tests pass and I ran the 'sat showrev' command on
spottedcow-sms1 and saw 'Sling' appear for the interconnects.