CRAYSAT-150: Add man page for 'sat hwinv'
Add a man page for 'sat hwinv'. Include all options in the man page.
Group options by list options and summary options. To avoid redundancy
refer to similar options for the node type for the other component
types. Add an examples section because there are a lot of options for
this subcommand, and it will be helpful to see examples of how they can
be used to achieve different views of the system hardware inventory.

Also remove an unused `xname` positional argument on the hwinv parser
because it is not currently implemented, and we don't want that
misleading info in the usage printed by argparse.

Revise description of list options argument group to not say tabular
format because the format is not tabular if `--format yaml` is
specified.

Fix a minor inconsistency in help for `--summary-<comp>-fields` option.

Test Description:
Built man page with `make all` in the `docs/man` directory and inspected
the resulting man page with `man ./sat-hwinv.8`. Ran all example
commands on spottedcow and verified that they behaved as described in
the examples section. Ran `sat hwinv -h` from my branch to ensure the
parser still works as expected.