CRAYSAT-165: Computed configuration defaults can be erroneous
Default for configuration option "username" when no configuration is
present should be the username of the account running sat, but this
is not relevant when the stock configuration file that ships with the
software is created. The value in this case should be empty, so that
it does not override a lower priority value that may be defined.

The solution here is to pass a callable object as the default, rather
than the string result of calling this object. The script under tools/
that generates the stock configuration file checks for a callable and
inserts a blank value, while the normally executing code calls if
callable, and uses the result.

Also, a mismatch between the command line argument "token_file" the
"token_filename" configuration option was resolved in favor of the
shorter form. This confusion led to an incorrect value in the option
spec, leading to the command line argument to be ignored!

Test Description:
the assert_defaults_set() method() was updated. all unit tests passing.
"SAT_CONFIG_FILE= sat auth" results in the correct username value, and
"tools/generate_default_config.py sat/config.py" has a blank value for
api_gateway.username.
The --token-file=foo command line option results in a token being writ-
ten to the indicated file. Also successfully tested the configuration
option analogously.