CRAYSAT-152: Format for all
Added a Report object to help unifiy tabular output. Integrated with
status and showrev subcommands, and hwinv --list-* options.

Adding yaml dump functionality to SatTable and integrating with showrev
and status.

sort-by argument defaults to 0, but now report won't sort if the
provided column wasn't found. This allows a user to do something like

    sat hwinv --sort-by None

And it won't sort.