CRAYSAT-191: Unittest report
Adding unit tests for report entry and yaml-readback.

This tests the sorting order and ports the deleted unittests from
sat status and added tests for the internal sorting of Report's
pretty table.

Also added the sub-string sort-by feature back in - which was removed
accidentally when the sat status sorting features were ported to the
Report class.