CRAYSAT-190: Linkhealth
Added linkhealth subcommand. This subcommand retrieves a list of router
xnames and uses Redfish to query the ports for their various aspects of
health before printing them into a table.

The Redfish credentials can be read from sat.toml. If they aren't
provided there, then the user will be prompted at command line.

TEST DESCRIPTION:

Mostly manual testing. Ran the tool repeatedly on Jolt1 - the only
machine with the right environment for solid e2e testing. All options
were exercised.

All existing unit tests pass, as well as the new ones introduced
for some functions within linkhealth.main and sat.redfish.

SIDE CHANGES:

Added logic to report to guarantee that dictionaries are stored
in the order of the headings. Also added a test to test_report to
verify that a defaultdict is considered valid.

Added dashes and numbers to the allowable character-set in filtering.py

Added description of how to configure a custom location for the
configuraiton file in sat.8.rst .

Fixing bug where an ImportError caused SAT to report that the containing
module did not exist when, in fact, it was a bad import within the
module.

Moving pester function from sat.diag.main to sat.util and moving
sat.cli.linkhealth.redfish to sat.redfish

Also added a message to get_username_and_pass to indicate which
username the password is for.

And fixed a bug with get_username_and_pass. What would happen is if
the username was blank in the conf file but the password was specified,
then the function would only prompt for a username. I thought this was
a misleading way for the function to behave.

KNOWN BUGS:

filtering on mixed types. Link speed, for example, can be ints or
'Not found'.