CRAYSAT-220: Rename `_subsequence` to `is_subsequence` and use in hwinv
Rename the function `_subsequence` to `is_subsequence`, so that its name
reflects the fact that it checks if one item is a subsequence of
another and that it is not a private function to this module. It is used
in two other locations. Use the function in the `ComponentField.matches`
method, so the way field names are specified in `sat hwinv` `--*-fields`
options is consistent with how query keys are specified in filters
specified with `--filter`.

Update man page to describe the subsequence behavior.

Test Description:
Tested the `--filter`, `--sort-by`, `--node-fields`, and
`--node-summary-fields` options with `sat hwinv` on spottedcow.