CRAYSAT-201: Refactor subcomand code structure.
This change moves all subcommands into a `cli` subpackage, and
changes the way in which it is loaded. All subcomands and argument
subparsers are now loaded dynamically, instead of by an `import`
statement. This improves performance, especially for tab-completion,
and allows for faster iteration, as fewer files need to be modified
when adding a new subcommand.

Test Description:
All subcommands were tested to ensure that no functionality was
affected by these changes. Unit tests were run to ensure no
regessions were introduced in automated testing.