CRAYSAT-113: Move hwinv components into shared `system` package
Move hwinv components into a shared `system` package since they will now
be used by both the `hwinv` and `hwmatch` subcommands. `system` may not
be the best name for this, and I don't like having a module named
`system` inside the `system` package, but that can be changed later. I'm
doing it this way for now to make it easier to rebase Erik's code for
`hwmatch` on top of this as he named the new package `system`. Perhaps a
better name for it would be `hw` or `hwcomponents`.

Test Description:
Tested that `sat hwinv` still worked.