CRAYSAT-151: Log warning messages with unused options
Log warning messages when `sat hwinv` receives unused options. E.g., if
`--node-fields` is specified without either `--list-all` or
`--list-nodes` being in effect, a warning message should be logged
indicating that `--node-fields` has no effect. As another example, if
`--node-summary-fields` or `--show-node-xnames` is specified without
`--summarize-all` or `--summarize-nodes` being in effect, a warning
message should be logged indicating that the option has no effect.

Change the default value for the `--show-<component>-xnames` option to
be `None`, so that we can tell whether the user specified the option or
not. With the default being a boolean, we can't tell whether the user
specified the option or not. This required changing where the value of
this option is used in the `ComponentSummary`.

Test Description:
Tested running `sat hwinv` with various combinations of options on
spottedcow. Verified that warning messages were logged at the end of the
output for any unused/irrelevant options. Also verified that the
functionality to turn on and off showing of xnames for each of the
different types of summaries still worked. Some option combos I tested:

*  --list-mems --node-fields xname
*  --list-nodes --node-enclosure-fields xname
*  --list-all --node-fields xname --node-enclosure-fields xname
*  --list-all --show-node-xnames off --show-mem-xnames on --show-proc-xnames
*  --summarize-all --node-summary-fields cab --mem-summary-fields cap
*  --node-fields xname --mem-fields xname --show-mem-xnames