CRAYSAT-195: Sort yaml output of Report
Sort the yaml output of the `Report.get_yaml` method, so that output is
sorted when the `--format yaml` and `--sort-by` options are used
together.

Take care of the sorting in a new `sort_data` method of `Report` which
can be called in both `get_yaml` and `get_pretty_table` rather than
sorting the `PrettyTable` instance using its `sortby` and `reversesort`
attributes. This allows sorting to be done the same way for both YAML
and PrettyTable output, which should mean fewer places to change code in
the future if something with sorting changes in the future.

Also fix a minor bug in the docstring of `filter_list` to indicate the
excpetions it raises and update the exception handling to handle all
exceptions it can raise.

Add a unit test to verify that YAML is sorted. It might be better and
more of a "unit" to test just the sorting and then mock the sorting step
out when calling the `get_yaml` and `get_pretty_table` methods, but this
is all right for now, and the coverage is the same.

Test Description:
Tested sorting of `sat status` and `sat showrev` output on spottedcow.
Tested with multiple different columns including columns that contained
output of heterogenous types that can't be compared without conversion
to a string. Tested both formats. Unit tests pass.