CRAYSAT-216: Remove necessary username/password from config file.
When running `sat diag`, it was previously necessary to record the
username and password for Redfish endpoints in the SAT config
file. This is obviously not ideal, as it presents a security
hole. This change makes it so that it is optional to record the
username and password in the `sat.toml` file; if users do not want to
save their password in plain text, they are prompted on each
invocation for the username and password for Redfish endpoints.

Test Description:
1. Run `sat diag` with various arguments on spottedcow to ensure no
   regressions were introduced.
2. Ensure unit tests pass.