CRAYSAT-203: Add Cray Proprietary license to LICENSE.txt
Add the Cray Proprietary license text to a LICENSE.txt file in the
top-level of our repository. This license was already specified in the
spec file, so there was no need to change that. Also add the license to
our setup.py, so the license information will appear when viewing our
installed package with `pip show`.

Test Description:
Installed the built rpm on spottedcow and verified "Cray Proprietary"
appears as the license both in pip and rpm.