CRAYSAT-218: Improve help messages for incorrect args in subcommands
This change makes it so that when an unrecognized argument is passed
to some subcommand on the command line, an appropriate help message is
printed to stderr based on which subcommand is in use. Previously, a
terse and relatively unhelpful error message was returned if an
unrecognized argument was passed. Now, passing an unrecognized
argument to some subcommand `sub` will have the same effect as running
`sat sub --help`.

Furthermore, this change improves error messages by describing how
exactly the given arguments are invalid. Error messages tell users if
they passed an invalid argument to the top level or to a subcommand,
and will alert users if no subcommand was used.

Test Description:

Ran `sat` with numerous subcommands and invalid options. Tested `sat`
with no arguments. Tested `sat` with no subcommand. Tested `sat` with
no subcommand and invalid arguments.