CRAYSAT-232: Improve `--xname` option documentation for `sat diag`
Improve the documentation of the `--xname` option in the usage
information for `sat diag`, and include an example of how to use it to
specify multiple xnames in the examples in the man page.

Test Description:
Built man page and viewed with `man` locally. Ran `sat diag -h` locally
and verified new usage information for `--xname`.