CRAYSAT-237: Present vs unconfigured
new linkhealth usage. All links should be reported, as well as
their presence in the table.

sat linkhealth used to determine whether it should report on a link based
on its presence in a static mapping provided by a configmap kubernetes
service.

That logic has now been replaced with a query to a redfish endpoint that
reports on the physical presence of a cable in the port, and the usage
and documentation has been updated.