CRAYSAT-221:
Parses the very heterogenous redfish telemetry sensor layout. Grabs
what was mentioned in the confluence page for the spike
https://connect.us.cray.com/confluence/display/XCCS/xtcheckhss+Spike+Findings

Sensor "status" and thresholds are not handled.

Floating point readings are truncated to four places after the decimal.
This is consistent with the actual precision in the values, for all sen-
sors of this type and other types that were observed.

Session capturing is implemented. This is intended to support "playback"
for testing purposes, but could be used for demonstration or diagnostics.

If the environment variable SAT_SENSORS_CAPTURE is set, the arguments for
"sat sensors" are augmented. Since they are conditional, autocompletion
will not work with them, but they do have usage help if argparse displays
it.

A temporary directory is created to contain the files generated. Upon
completion, the contents of this directory will be zipped into a .tgz
file, unless the option "--no-zip" is passed.

Test Description:

A unit test that loads captured sessions from the test resources,
replays them with capturing enabled, then compares the two captured
sessions.