CRAYSAT-297: Remove shasta logs index from exported mce-dashboard
Re-exported the MCE Dashboard without including related objects.
This omitted the shasta-logs index. Overwriting this index would
be a serious risk. The index is part of the Shasta Kibana and
does not belong to the MCE Dashboard.

Tested on surly2 where the MCE Dashboard had not been imported yet
(necessary for a clean test). Modified the search by changing a couple
of "and" operators to "or" operators. Observed NID counts and log
entries in the dashboard.