CRAYSAT-329: Fix 'sat status' handling of missing keys and xname sorting
Fix the `sat status` command so that it gracefully handles missing keys
in the data returned by requests to the HSM API. Also make sure that it
uses the `XName` class when adding the xnames to the table rows so that
they sort correctly.

Add unit tests for the cases of missing 'ID' and 'Flag' keys. Also add a
unit test to ensure the 'ID' value is converted to an `XName`.

Test Description:
Ran unit tests. Ran on surly2-ncn-w001 where the problem was encountered
and verified that it used "MISSING" for the "Flag" column on the node
that was missing that key. Also verified that the xnames were sorted
correctly now.