CRAYSAT-327: Add a changelog and increment the version number
Add a changelog in a CHANGELOG.md file that follows the format proposed
by the "Keep a Changelog" project.

Implement a python module in `tools` called `changelog` that implements
parsing the `CHANGELOG.md` file to extract the latest version. Then call
the function defined by this module to get the latest version from the
changelog inside `setup.py`. Then use `pkg_resources` to access this
version from within `sat.parser` so that `sat --version` will print the
version of `sat` that is installed.

Also get the version from the changelog in the spec file and set it in
the Version tag.

Test Description:
Tested a pip install in a virtual environment and verified that the
newly installed sat declared its version correctly in the output of `pip
show sat` and in the output of `sat --version`.

Tested an rpm build with the Jenkinsfile and verified that the rpm was
versioned correctly and that it printed the correct version when
installed through `pip show sat`, `sat --version`, and `rpm -qi
cray-sat` (tested on creek).