CRAYSAT-319: Fix filerting boolean precedence and revise man pages
Fixed problem with boolean precedence where a query such as
"A and B or C" did not match C.

Extended unit test with boolean query test that fails without
changes. Ran unit tests and all passed. Exercise queries on
real system including "A and B or C", "A or B and C",
"A and B", and singular comparison without boolean.

Updated man pages for filtering to include them in subcommands
with the filtering option. Reworded accordingly and made more
concise.

Tested building and viewing of man pages.