CRAYSAT-335: Fix `sat sensors` command
The `sensors` subcommand was still trying to do the splitting of
`args.xnames` on commas, but this is now handled in the argparse parser
group created in SAT_248, and `args.xnames` will either be `None` if no
xnames are specified or a list if xnames were specified.

Exit immediately in `do_sensors` if no xnames are provided because there
is nothing from which to get sensor readings. Update the man page to
describe this behavior, and create a unit test to test this.

Also update unit tests to pass in a list of xnames instead of a string
containing a comma-separated list of xnames.

Test Description:
Unit tests pass. Tested both with and without xnames specified on
surly2.

Built man page and inspected it with `man`.