CRAYSAT-327: Add contribution guidelines to README.md
Add guidelines for contributing to the project to `README.md` at the top
level of the repository. Include instructions on updating the changelog
in `CHANGELOG.md`, so that it stays up to date.

Test Description:
Pushed branch and examined markdown rendering in browser in bitbucket to
ensure it is formatted correctly.