CRAYSAT-289: Implement reading of release files in `sat showrev`
The file `/opt/cray/etc/release` that used to contain version
information about the products installed on a Shasta system was changed
to a directory that contains files for each product specifying version
information about that product. As a result `sat showrev --system` was
failing with a traceback.

Update the `sat showrev` command to parse the contents of any files it
finds under the directory `/opt/cray/etc/release`. Since these files
have multiple values in them, not just a single version number, and the
output of `--system` only has two columns in the table, move this output
to a new section of output controlled by a new `--products` option. Make
`--system` and `--products` the default behavior of `sat showrev` if no
options are specified.

Implement the logic for parsing the contents of the release directory in
a new `sat.cli.showrev.products` module, and write unit tests for this
file to achieve 100% coverage. Remove old code for parsing singlular
release file and unit tests for that old code.

Make some minor man page and usage information cleanups, e.g. to
consistently use the word 'display' rather than 'print'. Change exit
calls to sys.exit and reference a JIRA to fix early exits.

Test Description:
All unit tests pass with 100% coverage of new `products.py` module. Ran
`pycodestyle` and verified no style issues introduced. Copied to
thanos-ncn-w001 and ran `sat showrev` with various combinations of
options and verified the output matched what was in the file.