CRAYSAT-369: Remove tools python package from RPM
The `tools` python package in this repository is only needed at build
time, and is not intended to be included in the installation of this
code. Add `tools` to the `exclude` list of the call to `setup` in
`setup.py` in order to remove it. Also add a wildcard to exclude any
subpackages of the `tools` package that might be added in the future,
and add a wildcard to exclude any subpackages of the `tests` package,
which are currently erroneously being included in the RPM as well.

Test Description:
Pushed branch, downloaded RPM, and listed it to view its contents.
Verified that it only contained the sat python package and not the tests
or tools packages. Verified that the only removals from the files
included in the package were the tests and tools packages and all their
subpackages by comparing the listing of an older package to the listing
of the new package with `rpm -ql`.