CRAYSAT-252: Manpage examples consistency
Added "#" prompts where missing, changed "$" to "#".

Several instances of in-line example commentary were dedented to the
left margin.

Uniform preamble/introduction formatting.

As suggested in code review, all "preformatted" blocks indicated with
double colons are set with the double colons on their own lines with
blank lines above and below it. This includes one from the hwinv man
page that was not in the example section.

Feedback responses, mostly use of imperative and a bit more verbosity
where indicated.