CRAYSAT-382: Update the version in the changelog
Update the version in the changelog by incrementing the major version to
indicate added backwards-compatible functionality and bugfixes that have
been made since previous version.

Also add a line to the Fixed section for a bug that was fixed in the
sorting of xnames, and improve description of LBUG search query fix.
Also add a line to the Changed section for the Kibana dashboard column
consistency change.