CRAYSAT-291: Add ability to list drives in hwinv
Add the ability to list drives in `sat hwinv`. This is a new type of
component that has been recently added to HSM. Also add two new
properties (Drive Count and Total Drive Capacity) to the `Node` class
that can be displayed in a listing of nodes and can be used to summarize
the nodes.

Add a method to the `XName` class to get the `XName` of the parent node,
if possible. This method allows the code in the `relate_node_children`
to be generalized to work across Processors, MemoryModules, and Drives,
despite their different xname structure.

Update the man page for the newly added `--list-drives` option and its
associated `--drive-fields` option.

Add unit tests for the new `Drive` class, for the new `bytes_to_gib`
function in `util`, for the new `Drive` child type in the `Node` class,
and for new `get_node_parent` method in the `XName` class.

Add an example to the man page that shows how to get all the disks for a
specific node using filtering with wildcards.

Test Description:
Ran pycodestyle and verified no warnings. Ran unit tests with nosetests
and verified all tests pass. Built man page and inspected for correct
formatting.

Tested in a virtualenv on jolt1 as follows:
* Ran `sat hwinv --list-drives` with and without various `--drive-fields`
* Ran `sat hwinv --list-nodes` with `--node-fields` to view the various
  children fields that have been added or impacted by this change,
  including "drive", "memory", and "processor".
* Ran `sat hwinv --summarize-nodes --node-summary-fields drive` and
  verified that summaries worked with new drive fields.