CRAYSAT-302: Add listing of CMMRectifiers to `sat hwinv`
Add the ability to list CMMRectifier objects from HSM with the new
`--list-cmm-rectifiers` option to `sat hwinv`.

Add unit tests for new CMMRectifier class and update man page for new
options for `sat hwinv`.

Change the `location_info` property of the `BaseComponent` to return an
object of type `ComponentDataDict`, so that it will automatically
substitute the empty string with the `EMPTY_VALUE` and missing values
with the `MISSING_VALUE`. This is to support the `FirmwareVersion` of
the `CMMRectifier` objects in HSM, which appears in the `LocationInfo`
field, but is currently just empty. We'd like this to be hidden unless
the user specifies `--show-empty`.

Test Description:
All unit tests passing. pycodestyle reports no warnings. Built and
inspected `sat-hwinv` man page.

Tested against rocket-ncn-w001, which is the only place I've seen
CMMRectifier-type objects in HSM so far.