CRAYSAT-445: Clean up CHANGELOG and increment version to 2.0.0
Clean up CHANGELOG.md by fixing typos, cleaning up formatting, and
re-ordering the sections to match the order documented in the "keep a
changelog" convention.

Also move the description of the default behavior change for EMPTY and
MISSING columns to the first entry in the "Changed" section, since it is
probably one of the most impactful changes.

Since, at least two of the changes we've made here are somewhat
backwards-incompatible, increment the major version number, according to
the semantic versioning standard. The changes that may be backwards
incompatible are the removal of the `-s` option in `sat showrev`, and
the change to not show EMPTY and MISSING columns by default.

Test Description:
Inspected rendered CHANGELOG.md in Bitbucket to ensure that it renders
correctly. Ensured built package has the right version number.