CRAYSAT-272: Kibana dashboard for node heartbeat loss
Created Kibana search and Kibana dashboard that uses the search
(no visualization due to not having an xname count as xnames are
in message rather than hostname and there does not appear to be
a way to extract them within Kibana for visualization purposes).

Tested search and dashboard on rocket, plus importing of objects.