CRAYSAT-421: Linkhealth fixes
If "cable_present" is "Not Present", "MISSING" becomes "NOT APPLICABLE"
Added a reminder to the man page that if the cable is not plugged in,
some fields will not be defined.

It was determined that FlowControlConfiguration was removed from
Redfish intentionally, so it is removed from the reporting.

Ran nosetests
tested on thanos