CRAYSAT-467: Clean up changelog entries and increment version number
Clean up the entries in the changelog for this release of SAT, and then
set the version number according to semantic versioning guidelines.
Since we have added new items, and each change is backwards-compatible,
increment the minor version.

Cleanup includes:
* Fix heading level of "Removed" section
* Make section order consistent with instructions in README.md.
* Minor wording changes.
* Add mention of "MISSING" -> "NOT APPLICABLE" value change.

Test Description:
Pushed branch, examined rendered markdown in bitbucket, and checked that
the built rpm in artifactory had the new version number.