CRAYSAT-449: Display a message when no mismatches found
This will only display the message when 'pretty' format is specified,
with the idea that a tool that runs hwmatch could get confused when
trying to parse status text when it's expecting a YAML list.
Additionally, formats other than 'pretty' will continue to print empty
lists.

Test Description:
Installed this version of sat in a virtualenv on starlord-ncn-w001, then
ran sat hwmatch:
* with --format=yaml to verify the message is not displayed and
  an empty list is still given
* with --show-matches to verify the message is not displayed,
  because there is output
* with --format=pretty to verify the message is displayed
* with default arguments to verify the message is displayed