CRAYSAT-473: Update call to check-hsn-cables in `sat cablecheck`
Update `sat cablecheck` to call `/usr/bin/check-hsn-cables` instead of
`/usr/bin/check_hsn_cables.py` in response to the packaging changes made
in slingshot-tools.

Also collapse the two variables that were used before to refer to the
parent directory of the script and the script itself into a single
variable because we were not using them separately anywhere.

Test Description:
Unit tests pass. Ran in a virtual env on a shasta system with a test
script located at `/usr/bin/check-hsn-cables` to ensure that it was
invoked successfully. Double-checked path in latest slingshot-tools rpm
downloaded from artifactory.