CRAYSAT-448: Default to config vals for `no_headings` and `no_borders`
Default to the config values for the `no_headings` and `no_borders`
values of the `Report` class if they are not explicitly passed to the
constructor. These default config values are the values from the parsed
command-line args or from the config file, if not specified on the
command line. This makes it slightly easier to create a report without
having to call `get_config_val` and pass the result to the constructor
of the report.

Add unit tests for this new functionality.

Test Description:
Unit tests pass. Ran a couple commands (`status` and `hwinv`) with and
without `--no-headings` and `--no-borders` to confirm that they are
handled correctly still.