CRAYSAT-448: Implement first portion of `sat bootsys shutdown`
Implement the first portion of the `sat bootsys shutdown` command. This
portion of the command checks for any active sessions in the following
services on the Shasta system: BOS, CFS, CRUS, FUS/FAS, and NMD.

If any one of those services is found to have active services, the
command will exit with exit code 1. Information about each of the active
services is printed in a tabular format similar to other `sat` commands.
If none of those services are found to have active sessions, then the
command will instruct the user to continue with the shutdown process,
and then it will exit with exit code 0. In the future, it will continue
with the remainder of the automated shutdown procedure.

If any service fails to be queried for active sessions, the command will
indicate any active sessions it has found for the other services, and
then it will exit with exit code 1 and a message about the inability to
query the failed service(s) for active sessions. It will allow the user
to bypass this failure with the `--ignore-failed-services` option.

Add man page for `sat bootsys`.

Update CHANGELOG and increment version to prepare for merge to master.

Test Description:
Unit tests have been written and have achieved nearly 100% coverage on
the code in `service_activity`, and `main` modules in `sat.bootsys`
subpackage. Unit tests were also written for new functions added to
`util`, for new methods added to `FASClient` and `FUSClient`.

Tested manually against local FAS and on a couple real shasta systems
(starlord and mug). On mug, NMD, was not working, so that tested the
code for handling failed services. On starlord, all services were
working. Negated all conditions that check for active services to
exercise the code paths with active services.

Built and examined man page.