CRAYSAT-504: Fix URL for FAS in FASClient class
Fix the URL that we use to make requests to the Firmware Action Service
(FAS) by updating the `base_resource_path` in the FASClient class.

Also fix minor pycodestyle violation and update CHANGELOG.

Test Description:
Unit tests and pycodestyle pass.

Ran in a virtual environment on gamora, where FAS is installed and
operational and FUS is not operational. Verified with the following
commands:

* `sat firmware --snapshots`
* `sat firmware`
* `sat firmware -x x3000c0s1b0`
* `sat firmware -x x3000c0s1b0 -x x5000c1s0b0`
* `sat firmware --snapshots adn0`

Also verified that the expected requests were being made using
`--loglevel debug`.