CRAYSAT-482: Fix sat sensors man page
The sat-sensors man page incorrectly stated that xnames were required,
which is no longer the case. The default when no xnames are provided is
now to target all BMCs in the system, and this is mentioned earlier in
the man page. Also remove an incomplete sentence.

Test Description:
Built and inspected man page.