CRAYSAT-495: Implement Kubernetes pod status waiting.
This change adds code for waiting for Kubernetes pods to become available. This
change also adds a basic framework for creating simple "waiters", which block
the current thread until a given condition has been reached. This refactoring
is also used to re-implement all waiting code in the boot procedure. This
should greatly simplify the implementation and make it much more concise, while
abstracting away the specifics of waiting.

The waiting framework allows both synchronous and asynchronous waiting, so that
multiple conditions can be waited for at the same time. It is relatively
primitive and merely uses the threading module. The Waiter classes can also be
used as context managers. Within the block, other work can be done while
waiting for the manager's condition. Upon exiting the context, the thread will
block until the manager's condition has been reached.

This change also adds extensive unit tests for the new waiting framework and
concrete implementations.

A few additional bugs are also fixed which were encountered while doing
end-to-end testing on rocket:

* Ensure we actually start dhcpd by passing args.dry_run to
  `RunningService` constructor.
* Extend timeout for ssh to 10 minutes because it took longer than 2
  minutes for nodes to boot in testing. We really want to wait long enough
  here because we can't proceed with the ansible playbook anyway if nodes
  are not SSH-able.
* Raise a SystemExit if any nodes do time out waiting to be accessible.
  In the future, we may want to allow for prompting the user whether they
  want to keep waiting or not. Or maybe just wait indefinitely and allow
  the user to quit with Ctrl-C or something.
* Wait for the kubernetes API to become available before querying it.

Test Description:

Added unit tests to cover the bulk of added functionality. In addition, an
end-to-end test of booting was performed on rocket which covered much of the
shutdown and boot process using the updated shutdown and boot code.