CRAYSAT-495: Refactor GroupWaiter for individual waiter
This change factors out a Waiter class which only waits for a single condition.
This is a general case of the GroupWaiter class but it is a simpler
implementation better suited for waiting for e.g. Ceph health status. Unit
tests have also been added for the new functionality.

Test Description:
Added unit tests for new Waiter class. No end-to-end tests have been performed
yet.