CRAYSAT-491: Fix bugs in sat.cli.bootsys.bos
* Fix a bug with pluralization of "session templates" in a debug log
  message.
* Make sure that the `handle_state_check_failure` function always returns
  the list that it is supposed to, both in the recursive case and in the
  skip case. Otherwise an exception occurs when trying to log the session
  templates that will have a shutdown operation performed on them because
  they are `None`.
* When no session templates are found that need to have the shutdown
  operation performed, do not even call `do_parallel_shutdowns`.

Test Description:
Tested the log statement in isolation in a python interpreter to make
sure inflect usage and format args are correct.

Tested the second fix on rocket-ncn-w001 when state check was failing.

Did not test the third change, but it's small, and no inspection
warnings in pycharm or style violations in pycodestyle.