CRAYSAT-496: Add implementation of Ceph waiter
This change adds an implementation of a Waiter for the Ceph cluster, which
waits for it to have a HEALTH_OK status before proceeding. This implements step
4.1.11 in the Shasta Admin Guide. This change also adds unit tests for this new
functionality.

Test Description:

Unit tests have been written for the waiter, and some initial testing
was done on mug. Only the positive case was tested.