CRAYSAT-491: Run platform-shutdown.yml ansible playbook during shutdown
Run the `platform-shutdown.yml` ansible playbook during the `sat bootsys
shutdown` command after performing the BOS shutdown of computes and UANs
and before performing the shutdown and power off of the management NCNs.

Note that this depends on SAT_535, which added the "prompt" tag to the
`platform-shutdown.yml` ansible playbook and on SAT_536, which makes the
ansible playbook timeout appropriately on the `crictl stop` task, so
that it doesn't hang indefinitely.

Update the unit tests to get them to pass and to do assertions on the
way `do_shutdown_playbook` is called in `do_shutdown`.

Test Description:
This has been tested once on rocket-ncn-w001 to shut that system down.
All unit tests are passing, and pycodestyle reports no style violations.