CRAYSAT-554: Fix usage of logging.warning
Fix an instance where we were using `logging.warning` instead of
`LOGGER.warning` where `LOGGER` is the module-level logger. This was
resulting in `logging.basicConfig` being called because no handlers are
defined on the root logger. This added a handler to the root logger with
level NOTSET, so all messages were logged by that handler. This resulted
in log messages being printed to stderr multiple times, and it was also
logging debug-level messages to stderr even when the configured log
level was higher than debug.

Test Description:
Tested on gamora with `sat showrev --system` which was logging errors,
warnings, and debug messages. Tested with different config values for
`logging.file_level` and `logging.stderr_level`. Tested with
`--loglevel` option that overrides both of these. Verified that only the
correct messages were logged.