CRAYSAT-465: Boot computes and UANs with BOS in `sat bootsys boot`
Add the boot of compute nodes and User Access Nodes (UANs) using the
Boot Orchestration Service (BOS) to the `sat bootsys boot` command.

This generalizes the implementation of the shutdown using BOS, so that
the same code can be reused for the boot using BOS.

Implement a couple unit tests for the `do_boot` function and remove the
`skip` decorator.

Test Description:
Tested on mug with a test script that calls `do_bos_operations` with an
operation given on the command line. Note that the BOS boot sessions
failed due to a problem on mug, but the shutdown sessions succeeded, and
the code worked correctly and reported success or failure.  All unit
tests pass and pycodestyle reports no violations.