CRAYSAT-497: Add implementation of BGP peering waiter.
This change adds a completed implementation with tests for the BGP spine
peering waiter. This causes the waiter to reset the BGP routes before starting
if it detects that they are idle. A small refactor was done to the
Waiter/GroupWaiter classes (a method was moved from the subclass GroupWaiter to
the superclass Waiter) in order to support this change.

This change also adds unit tests for the added functionality.

Test Description:

This change was not tested on live hardware as it could possibly be disruptive,
thus it has only been tested as of yet with unit tests.