CRAYSAT-465: Fix storage of BOS session ID and refactor
Fix the storage of the BOS session ID to properly extract the session ID
from the `href` attribute of the `link` attribute returned by the
request to create the BOS session.

Refactor the code to just store the status about the BOSSessionThread in
the object itself as attributes rather than in a separate dictionary.
The separate dictionary really isn't necessary because the code that
creates the thread saves the thread object and can access its attributes
to query its status.

Add a unit test for the happy path through the `create_session` method
of `BOSSessionThread` to test the `posixpath.dirname` call.

Test Description:
Tested on mug-ncn-w001 with a test script that calls
`do_bos_operations` with a given operation. Tested with both 'boot' and
'shutdown' operations and verified that they correctly wait for the
sessions to complete. Note that the BOS sessions failed on mug due to
software problems on that system, but this code is still pretty well
exercised.

All unit tests pass. pycodestyle reports no violations.