CRAYSAT-498: Add implementation of HSN bringup waiter.
This change adds an implementation of a Waiter for the HSN
bringup. This runs the
/opt/cray/crayctl/ansible_framework/main/ncmp_hsn_bringup.yaml
playbook and then uses the /opt/cray/bringup-fabric/status.sh script
to monitor the progress of the HSN bringup.

Test Description:
This has not been tested on real hardware, though some basic unit
tests have been written to cover the new functionality.